# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deloitte Initiative Discovery Platform",
    "description": "\n    ## Deloitte Initiative Discovery Platform API\n    \n    A unified platform for Deloitte analysts to discover and engage with firm initiatives.\n    \n    ### Features:\n    * **Authentication**: Azure AD B2C SSO integration\n    * **User Profiles**: Manage analyst and leader profiles\n    * **Initiative Management**: Create, update, and manage initiatives\n    * **Discovery Tools**: Search, filter, and browse initiatives\n    * **AI Recommendations**: Personalized initiative matching\n    * **Engagement**: Save, bookmark, and track initiatives\n    \n    ### User Roles:\n    * **Analyst**: Browse, search, and receive recommendations\n    * **Leader**: Create and manage initiatives\n    ",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login user",
        "description": "Login endpoint (simplified version).\n\nIn production, this would:\n1. Validate Azure AD B2C token\n2. Extract user info from token\n3. Create or update user in database\n4. Return JWT for API access\n\nFor now, it creates a user if doesn't exist and returns JWT.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register new user",
        "description": "Register a new user.\n\nIn production, this would be replaced by automatic provisioning from Azure AD.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/azure-login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Azure AD B2C login (future)",
        "description": "Azure AD B2C login endpoint (placeholder for production implementation).\n\nThis endpoint would:\n1. Validate the Azure AD token using MSAL\n2. Extract user claims (email, name, groups)\n3. Map Azure AD groups to user roles\n4. Create/update user in database\n5. Return JWT for API access",
        "operationId": "azure_login_api_v1_auth_azure_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AzureADToken"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/engagement/save": {
      "post": {
        "tags": [
          "Engagement"
        ],
        "summary": "Save/bookmark initiative",
        "description": "Save/bookmark an initiative for later.\n\nSaved initiatives appear in the user's saved list.",
        "operationId": "save_initiative_api_v1_engagement_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveInitiativeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/engagement/save/{initiative_id}": {
      "delete": {
        "tags": [
          "Engagement"
        ],
        "summary": "Remove saved initiative",
        "description": "Remove an initiative from saved/bookmarked list.",
        "operationId": "unsave_initiative_api_v1_engagement_save__initiative_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "initiative_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Initiative Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/engagement/saved": {
      "get": {
        "tags": [
          "Engagement"
        ],
        "summary": "Get saved initiatives",
        "description": "Get all initiatives saved/bookmarked by the current user.",
        "operationId": "get_saved_initiatives_api_v1_engagement_saved_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InitiativeResponse"
                  },
                  "type": "array",
                  "title": "Response Get Saved Initiatives Api V1 Engagement Saved Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/engagement/apply": {
      "post": {
        "tags": [
          "Engagement"
        ],
        "summary": "Apply to initiative",
        "description": "Express interest or apply to an initiative.\n\nThe application includes an optional message to the initiative owner.",
        "operationId": "apply_to_initiative_api_v1_engagement_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/engagement/applications": {
      "get": {
        "tags": [
          "Engagement"
        ],
        "summary": "Get my applications",
        "description": "Get all applications submitted by the current user.",
        "operationId": "get_my_applications_api_v1_engagement_applications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApplicationResponse"
                  },
                  "type": "array",
                  "title": "Response Get My Applications Api V1 Engagement Applications Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/engagement/initiative/{initiative_id}/applications": {
      "get": {
        "tags": [
          "Engagement"
        ],
        "summary": "Get initiative applications",
        "description": "Get all applications for a specific initiative.\n\nOnly accessible by the initiative owner or admin.",
        "operationId": "get_initiative_applications_api_v1_engagement_initiative__initiative_id__applications_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "initiative_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Initiative Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationResponse"
                  },
                  "title": "Response Get Initiative Applications Api V1 Engagement Initiative  Initiative Id  Applications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplicationRequest": {
        "properties": {
          "initiative_id": {
            "type": "integer",
            "title": "Initiative Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "initiative_id"
        ],
        "title": "ApplicationRequest"
      },
      "ApplicationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "initiative_id": {
            "type": "integer",
            "title": "Initiative Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "applied_at": {
            "type": "string",
            "format": "date-time",
            "title": "Applied At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "initiative_id",
          "message",
          "applied_at",
          "status"
        ],
        "title": "ApplicationResponse"
      },
      "AzureADToken": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "AzureADToken",
        "description": "Azure AD token (for future implementation)"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InitiativeDuration": {
        "type": "string",
        "enum": [
          "short_term",
          "ongoing",
          "fixed_duration"
        ],
        "title": "InitiativeDuration"
      },
...
      "InitiativeResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "practice_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Practice Area"
          },
          "skills_needed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Needed"
          },
          "industries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Industries"
          },
          "time_commitment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Commitment"
          },
          "duration": {
            "allOf": [
              {
                "$ref": "#/components/schemas/InitiativeDuration"
              }
            ],
            "default": "ongoing"
          },
          "duration_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Details"
          },
          "role_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Type"
          },
          "contact_person": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Person"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/InitiativeStatus"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "owner_id": {
            "type": "integer",
            "title": "Owner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "view_count": {
            "type": "integer",
            "title": "View Count",
            "default": 0
          },
          "save_count": {
            "type": "integer",
            "title": "Save Count",
            "default": 0
          },
          "application_count": {
            "type": "integer",
            "title": "Application Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "id",
          "status",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "InitiativeResponse"
      },
      "InitiativeStatus": {
        "type": "string",
        "enum": [
          "open",
          "active",
          "full",
          "closed"
        ],
        "title": "InitiativeStatus"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "LoginRequest",
        "description": "Login request with email (simulating Azure AD)"
      },
...
      "SaveInitiativeRequest": {
        "properties": {
          "initiative_id": {
            "type": "integer",
            "title": "Initiative Id"
          }
        },
        "type": "object",
        "required": [
          "initiative_id"
        ],
        "title": "SaveInitiativeRequest"
      },
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "title": "Token"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ],
            "default": "analyst"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name"
        ],
        "title": "UserCreate"
      },
      "UserResponse": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "practice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Practice"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "interests": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Interests"
          },
          "industries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Industries"
          },
          "experience_years": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Years"
          },
          "certifications": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Certifications"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "last_login": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name",
          "id",
          "role",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "analyst",
          "leader",
          "admin"
        ],
        "title": "UserRole"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}